Fixed what I broke
So it looks like when you adjust the asciiz text in the .data segment,
it reallllly messes with the later alignment of data, and results in
errors when trying to load the size of the image from the header. Spent
about an hour trying to fix it, but I'm not sure how to align
everything.