Added output file check/
Added two things:
1.) The first is a quick message to tell the user that the filtering
might take a while, so chill.
2.) A catch block to account for the case when you want to output to a
file that can't be written to (such as a directory that doesn't exist).
Printing this error message results in MARS spewing crap text, and the
program becoming unresponsive. Not sure why this is happening, but it
occurs if I try other messages as well.